backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
snowflake-connector-python==3.17.3
python-dotenv==1.1.1

